Implement some of landing page, add description to projects

+++ BlockBot.Web/Controllers/ProjectsController.cs
@@ -1,7 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Text.Encodings.Web;
 using System.Threading.Tasks;
+using BlockBot.Module.Aws.Models;
+using BlockBot.Module.Aws.ServiceInterfaces;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
@@ -11,11 +14,19 @@ namespace BlockBot.Web.Controllers
 {
     public class ProjectsController : Controller
     {
+        private readonly IApiGatewayService _apiGatewayService;
         private readonly ApplicationDbContext _context;
+        private readonly UrlEncoder _urlEncoder;
+        private readonly ApplicationUserManager _userManager;
 
-        public ProjectsController(ApplicationDbContext context)
-        {
+        public ProjectsController(ApplicationDbContext context,
+            ApplicationUserManager userManager,
+            UrlEncoder urlEncoder,
+            IApiGatewayService apiGatewayService){
             _context = context;
+            _userManager = userManager;
+            _urlEncoder = urlEncoder;
+            _apiGatewayService = apiGatewayService;
         }
 
         // GET: Projects/Details/5
@@ -34,13 +45,22 @@ namespace BlockBot.Web.Controllers
                 return NotFound();
             }
 
+            ApplicationUser user = await _userManager.GetUserAsync(User);
+            if (user == null)
+            {
+                return NotFound($"Unable to load user with ID '{_userManager.GetUserId(User)}'.");
+            }
+            if (project.OwnerId != user.Id)
+            {
+                return Unauthorized();
+            }
+
             return View(project);
         }
 
         // GET: Projects/Create
         public IActionResult Create()
         {
-            ViewData["OwnerId"] = new SelectList(_context.Users, "Id", "Id");
             return View();
         }
 
@@ -49,16 +69,29 @@ namespace BlockBot.Web.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Name,IsPublic")] Project project)
+        public async Task<IActionResult> Create([Bind("Name,IsPublic,Description")] Project project)
         {
             if (ModelState.IsValid)
             {
+                ApplicationUser user = await _userManager.GetUserAsync(User);
+                if (user == null)
+                {
+                    return NotFound($"Unable to load user with ID '{_userManager.GetUserId(User)}'.");
+                }
+                project.OwnerId = user.Id;
+                project.XML =
+                    "<xml xmlns=\"http://www.w3.org/1999/xhtml\" id=\"workspaceBlocks\" style=\"display: none;\"></xml>";
+                ApiGatewayRestApi result = await _apiGatewayService.CreateApiGateway(
+                    _urlEncoder.Encode(user.NormalizedUserName + "-" + project.Name).ToLowerInvariant(),
+                    "TODO add permalink to project");
+                project.RestApiId = result.RestApiId;
+
                 project.Id = Guid.NewGuid();
                 _context.Add(project);
                 await _context.SaveChangesAsync();
                 return RedirectToAction("Dashboard", "Dashboard");
             }
-            ViewData["OwnerId"] = new SelectList(_context.Users, "Id", "Id", project.OwnerId);
+
             return View(project);
         }
 
@@ -75,7 +108,17 @@ namespace BlockBot.Web.Controllers
             {
                 return NotFound();
             }
-            ViewData["OwnerId"] = new SelectList(_context.Users, "Id", "Id", project.OwnerId);
+            
+            ApplicationUser user = await _userManager.GetUserAsync(User);
+            if (user == null)
+            {
+                return NotFound($"Unable to load user with ID '{_userManager.GetUserId(User)}'.");
+            }
+            if (project.OwnerId != user.Id)
+            {
+                return Unauthorized();
+            }
+
             return View(project);
         }
 
@@ -84,18 +127,36 @@ namespace BlockBot.Web.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(Guid id, [Bind("Name,IsPublic")] Project project)
+        public async Task<IActionResult> Edit(Guid id, [Bind("Id,Name,Description,IsPublic")] Project project)
         {
             if (id != project.Id)
             {
                 return NotFound();
             }
 
+            ApplicationUser user = await _userManager.GetUserAsync(User);
+            if (user == null)
+            {
+                return NotFound($"Unable to load user with ID '{_userManager.GetUserId(User)}'.");
+            }
+
+            Project originalProject = await _context.Projects.FindAsync(id);
+            if (originalProject == null)
+            {
+                return NotFound();
+            }
+            if (originalProject.OwnerId != user.Id)
+            {
+                return Unauthorized();
+            }
+
             if (ModelState.IsValid)
             {
                 try
                 {
-                    _context.Update(project);
+                    originalProject.Name = project.Name;
+                    originalProject.Description = project.Description;
+                    originalProject.IsPublic = project.IsPublic;
                     await _context.SaveChangesAsync();
                 }
                 catch (DbUpdateConcurrencyException)
@@ -111,7 +172,7 @@ namespace BlockBot.Web.Controllers
                 }
                 return RedirectToAction("Dashboard", "Dashboard");
             }
-            ViewData["OwnerId"] = new SelectList(_context.Users, "Id", "Id", project.OwnerId);
+
             return View(project);
         }
 
@@ -131,6 +192,16 @@ namespace BlockBot.Web.Controllers
                 return NotFound();
             }
 
+            ApplicationUser user = await _userManager.GetUserAsync(User);
+            if (user == null)
+            {
+                return NotFound($"Unable to load user with ID '{_userManager.GetUserId(User)}'.");
+            }
+            if (project.OwnerId != user.Id)
+            {
+                return Unauthorized();
+            }
+
             return View(project);
         }
 
@@ -140,6 +211,25 @@ namespace BlockBot.Web.Controllers
         public async Task<IActionResult> DeleteConfirmed(Guid id)
         {
             var project = await _context.Projects.FindAsync(id);
+
+            if (project == null)
+            {
+                return NotFound();
+            }
+
+            ApplicationUser user = await _userManager.GetUserAsync(User);
+            if (user == null)
+            {
+                return NotFound($"Unable to load user with ID '{_userManager.GetUserId(User)}'.");
+            }
+            if (project.OwnerId != user.Id)
+            {
+                return Unauthorized();
+            }
+            // exception should be thrown if delete fails
+            await _apiGatewayService.DeleteApiGateway(project.RestApiId);
+            // TODO delete S3 and Lambda resources, Twilio resources
+
             _context.Projects.Remove(project);
             await _context.SaveChangesAsync();
             return RedirectToAction("Dashboard", "Dashboard");
